+++ tasmota/tasmota_template.h
@@ -236,6 +236,8 @@ enum UserSelectablePins {
   GPIO_BL0940_RX,      // BL0940 serial interface
   GPIO_TELEINFO_RX,      // BL0940 serial interface
   GPIO_TELEINFO_ENABLE, // BL0940 serial interface
+  GPIO_TCP_TX,         // TCP Serial bridge
+  GPIO_TCP_RX,         // TCP Serial bridge
   GPIO_SENSOR_END };
 
 // Programmer selectable GPIO functionality
@@ -327,7 +329,8 @@ const char kSensorNames[] PROGMEM =
   D_SENSOR_BOILER_OT_RX "|" D_SENSOR_BOILER_OT_TX "|"
   D_SENSOR_WINDMETER_SPEED "|"
   D_SENSOR_BL0940_RX  "|"
-  D_SENSOR_TELEINFO_RX "|" D_SENSOR_TELEINFO_ENABLE
+  D_SENSOR_TELEINFO_RX "|" D_SENSOR_TELEINFO_ENABLE "|"
+  D_SENSOR_TCP_TXD "|" D_SENSOR_TCP_RXD
   ;
 
 const char kSensorNamesFixed[] PROGMEM =
@@ -689,7 +692,10 @@ const uint8_t kGpioNiceList[] PROGMEM = {
   GPIO_TELEINFO_RX,
   GPIO_TELEINFO_ENABLE,
 #endif
-
+#ifdef USE_TCP_BRIDGE
+  AGPIO(GPIO_TCP_TX),      // TCP Serial bridge
+  AGPIO(GPIO_TCP_RX),      // TCP Serial bridge
+#endif
 };
 
 /********************************************************************************************/